Restructed data

+++ Assets/Main/Map/Panels/Events/Scripts/Examples/HungerEvent.cs
@@ -76,7 +76,7 @@ public struct HungerEvent : I_Event {
 
     public bool Migration() {
         info.EventResult(settings.Localization.Events[0].Results[1].Info);
-        map.data.GetRandomCivilizationID(region, out int civID);
+        map.data.GetRandomCivilizationID(region, out float civID);
         migration.CreateMigration(region, MigrateRegion, civID);
         return true;
     }
